+++ budgie.go
@@ -5,32 +5,7 @@ import (
 	"os"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/charmbracelet/lipgloss"
 )
-
-var textStyle = lipgloss.NewStyle().
-	Bold(true).
-	Foreground(lipgloss.Color("#FAFAFA")).
-	Background(lipgloss.Color("#1f4fd1")).
-	Align(lipgloss.Left)
-
-var promptStyle = lipgloss.NewStyle().
-	Bold(true).
-	Foreground(lipgloss.Color("#FAFAFA")).
-	Background(lipgloss.Color("#fcc319")).
-	Align(lipgloss.Left)
-
-var selectedStyle = lipgloss.NewStyle().
-	Bold(true).
-	Foreground(lipgloss.Color("#FAFAFA")).
-	Background(lipgloss.Color("#32d147")).
-	Align(lipgloss.Left)
-
-var errorStyle = lipgloss.NewStyle().
-	Bold(true).
-	Foreground(lipgloss.Color("#FAFAFA")).
-	Background(lipgloss.Color("#f542c2")).
-	Align(lipgloss.Left)
 
 const MongoDb = "budgie"
 const MongoCollection = "expenses"
