+++ home_screen.go
@@ -76,7 +76,6 @@ func (m homeScreenModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 				delete(m.selected, m.cursor)
 			} else {
 				m.selected[m.cursor] = struct{}{}
-				fmt.Println(m.selected)
 			}
 		}
 	}
@@ -100,7 +99,7 @@ func (m homeScreenModel) View() string {
 			cursor = ">" // cursor!
 			content += selectedStyle.Width(HomeScreenWidth).Render(fmt.Sprintf("%s %s", cursor, choice))
 		} else {
-			content += promptStyle.Width(HomeScreenWidth).Render(fmt.Sprintf("%s %s", cursor, choice))
+			content += errorStyle.Width(HomeScreenWidth).Render(fmt.Sprintf("%s %s", cursor, choice))
 		}
 
 		if i < len(m.choices)-1 {
