+++ styles.go
@@ -1,0 +1,27 @@
+package main
+
+import "github.com/charmbracelet/lipgloss"
+
+var textStyle = lipgloss.NewStyle().
+	Bold(true).
+	Foreground(lipgloss.Color("#FAFAFA")).
+	Background(lipgloss.Color("#1f4fd1")).
+	Align(lipgloss.Left)
+
+var promptStyle = lipgloss.NewStyle().
+	Bold(true).
+	Foreground(lipgloss.Color("#FAFAFA")).
+	Background(lipgloss.Color("#fcc319")).
+	Align(lipgloss.Left)
+
+var selectedStyle = lipgloss.NewStyle().
+	Bold(true).
+	Foreground(lipgloss.Color("#FAFAFA")).
+	Background(lipgloss.Color("#32d147")).
+	Align(lipgloss.Left)
+
+var errorStyle = lipgloss.NewStyle().
+	Bold(true).
+	Foreground(lipgloss.Color("#FAFAFA")).
+	Background(lipgloss.Color("#f542c2")).
+	Align(lipgloss.Left)
